+++ CHANGELOG.md
@@ -16,8 +16,10 @@ ADDED:
 
 CHANGED:
 
-- update of com.github.ben-manes.versions.gradle.plugin to new version 0.48.0
-- update of gradle-plugin dependency 'com.diffplug.spotless:spotless-plugin-gradle' to new minor version 6.21.0
+- update gradle to new version 8.7
+- update of com.github.ben-manes.versions.gradle.plugin to new version 0.51.0
+- update of gradle-plugin dependency 'org.ajoberstar.grgit:grgit-gradle' in version 5.2.2
+- update of gradle-plugin dependency 'com.diffplug.spotless:spotless-plugin-gradle' to new minor version 6.25.0
 
 Version 27
 -------------